matplot

+++ numpys/arraymore/DemoSwapByte.py
@@ -2,8 +2,8 @@
 
 import numpy as num
 
-hello=num.array([12,5,90,10,0,34],dtype=num.intc)
+hello=num.array([12,5,90,10,0,34],dtype=num.int16)
 
 print(hello)
 
-print(map(hex,hello))
+print (map(hex,hello))
